Moved KeyEntryIdConverter to Converters folder
