+++ KeyManager.Library.UI/PublishKeyStoreDialog.xaml
@@ -7,12 +7,13 @@
              xmlns:local="clr-namespace:Leosac.KeyManager.Library.UI"
              xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
              xmlns:properties="clr-namespace:Leosac.KeyManager.Library.UI.Properties"
+             xmlns:converters="clr-namespace:Leosac.KeyManager.Library.UI.Converters"
              mc:Ignorable="d"
              d:DataContext="{d:DesignInstance domain:PublishKeyStoreDialogViewModel}"
              d:DesignHeight="200" d:DesignWidth="500"
              MinWidth="300">
     <UserControl.Resources>
-        <local:KeyEntryIdConverter x:Key="KeyEntryIdConverter"/>
+        <converters:KeyEntryIdConverter x:Key="KeyEntryIdConverter"/>
         <materialDesign:BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter" />
     </UserControl.Resources>
     <DockPanel LastChildFill="True">
